HarvesterFactory: use Laminas Proxy adapter to allow connections over a proxy
Sometimes the harvesting service works behind an HTTP proxy and
therefore need to be able to communicate with the outer world trough it.
Laminas provides a dedicated Laminas\Http\Client\Adapter\Proxy adapter
for it. If no proxy settings are present, this adapter falls back to a
default Socket adapter so there is no harm in setting it up for all
connections.